third_party/WebKit/LayoutTests/http/tests/devtools/sources/autocomplete-hide-on-smart-brace-expected.txt

Verify that suggest box gets hidden whenever a cursor jumps over smart brace.


Running: testSummonSuggestBox
Suggestions shown.

Running: testTypeSmartBrace
Suggestions hidden.
